+++ app/src/test/java/reikai/presentation/reader/ReaderProviderConformanceTest.kt
@@ -101,5 +101,6 @@ class NovelOrientationProbe : ReaderOrientationProbe {
                 )
             },
             novelPreferences = NovelPreferences(InMemoryPreferenceStore()),
+            fontManager = mockk(),
         )
 }